+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Automatically create the prototypal wrapper methods for the function inheritance
 * `Function#setProperty` now accepts a single object to set multiple properties
+* Add `Stream.Delayed()` class
 
 ## 0.7.0 (2020-03-13)